pyproject.toml
[tool.poetry]
name = "kanade-bot"
version = "0.1.0"
description = "基于绪山真寻bot二次创作而来的一只可爱的宵崎奏同人bot"
authors = ["cYanosora <1362747420@qq.com>"]
license = "AGPL"
readme = "README.md"
packages = [{include = "kanade_bot"}]

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0rc1"
nonebot-adapter-onebot = "^2.1.5"
nonebot-plugin-apscheduler = "^0.2.0"
gino = "^1.0.1"
httpx = "^0.22.0"
matplotlib = "^3.5.1"
ujson = "^5.1.0"
aiocache = "^0.11.1"
pypinyin = "^0.46.0"
"ruamel.yaml" = "^0.17.21"
psutil = "^5.9.0"
playwright = "^1.18.2"
ImageHash = "^4.2.1"
emoji = "^1.7.0"
jieba = "^0.42.1"
lxml = "4.6.5"
mutagen = "^1.45.1"
jinja2 = "^3.1.1"
requests = "^2.28.1"
loguru = "^0.6.0"
fonttools = "^4.0.0"
Pillow = "^9.0.1"
numpy = "^1.20.0"
opencv-python-headless = "^4.0.0"
bbcode = "^1.1.0"
aiofiles = "^0.8.0"
aiohttp = "^3.8.3"
imageio = "^2.23.0"
retrying = "^1.3.4"
markdown = "^3.4.1"
zhconv = "^1.4.3"
rich = "^12.6.0"
pinyin = "^0.4.0"
bs4 = "^0.0.1"
thefuzz = "^0.19.0"
baike = "^1.5.0"
nonebot-plugin-htmlrender = "^0.2.0.1"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
